server: add comments
